+++ internal/lefthook/runner/executor.go
@@ -16,5 +16,5 @@ type ExecuteOptions struct {
 // It is used here for testing purpose mostly.
 type Executor interface {
 	Execute(opts ExecuteOptions) (*bytes.Buffer, error)
-	RawExecute(command string, args ...string) error
+	RawExecute(command string, args ...string) (*bytes.Buffer, error)
 }